move Configuration to Katello namespace
split configuration to files by class
move logging to katello dir
