+++ src/Gemfile32
@@ -7,7 +7,7 @@ end
 # load Katello configuration
 path = File.expand_path('../lib', __FILE__)
 $LOAD_PATH << path unless $LOAD_PATH.include? path
-require 'katello_config'
+require 'katello/load_configuration'
 
 # When adding new version requirement check out EPEL6 repository first
 # and use this version if possible. Also check Fedora version (usually higher).
