+++ src/lib/katello/logging.rb
