+++ src/script/rails
@@ -12,7 +12,7 @@ require 'yaml'
 # load Katello configuration
 path = File.expand_path('../../lib', __FILE__)
 $LOAD_PATH << path unless $LOAD_PATH.include? path
-require 'katello_config'
+require 'katello/load_configuration'
 
 #
 # this is to enable ssl and ssl cert-based client authentication
